% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Build a map where the events (accidents) occurs}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{this}{function take two arguments, first the \code{state.num} argument, which speciies
the state which we are interested.}

\item{the}{second argument is the \code{year} argument which specifies the year which we are interested.}
}
\value{
this function returns a map where we can see how many accidents occurs and where they occurs.
If the input state number doesn't exist in the data the function running stops
and the message "invalid STATE number: " is returned
If there is not data for specified state number and year
the message "no accidents to plot" is returned.
}
\description{
This function converts a CSV file in a dataframe and select a specified year and state number,
then using this dataframe builds a map in order to indicate where the accidents occurs and how many accidents occurs in one place.
}
\examples{
\dontrun{fars_map_state(c(18,2013))}
\dontrun{fars_map_state(c(18,2018))}
\dontrun{fars_map_state(c(82:2013))}

}
