% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Convert a CSV file in a dataframe, in order to manage it in R.}
\usage{
fars_read(filename)
}
\arguments{
\item{The}{only parameter of this function is the name of a CSV file which we want to read in R.
(using the \code{filename} argument).}
}
\value{
This function returns a dataframe created from a CSV data file.
This function stops and returns a message if the file name does not exist.
}
\description{
This function can be used to convert a CSV file in a dataframe, in order
to manage it in R.
This function stops and returns a message if the file name does not exist.
}
\examples{
\dontrun{fars_read("data/accident_2013.csv.bz2")}
\dontrun{fars_read("data/accident_2018.csv.bz2")}

}
