% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{This}{function takes only one argument a vector of several years for the selected years,
(using the \code{years} argument).}
}
\value{
This function returns a dataframe, which summarizes the number of events
(accidents) per year.
}
\description{
This function allows to summarize the number of events (accidents) per months and year.
First it reads the CSV file of the selected years and convert it in a dataframe
using the function above: fars_read_years.
Then it binds the columns and group the lines per month and years using the package deplyr.
Finally, it summarize the number of events (accidents) per month and years using the package deplyr.
And reorganize the dataframe using the function spread of the package tidyr.
}
\examples{
\dontrun{fars_summarize_years(c(2013:2015))}
\dontrun{fars_summarize_years(c(2015:2018))}

}
