% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Build filename with the year}
\usage{
make_filename(year)
}
\arguments{
\item{This}{function takes only one argument a 4 digits number for the year,
(using the \code{year} argument).}
}
\value{
This function returns simply the name of the CSV file, which we want
to convert in dataframe with the function above fars_read.
}
\description{
This filename will be used in the function above fars_read.
}
\examples{
\dontrun{make_filename(2013)}
\dontrun{make_filename(2018)}

}
